ajout création ouvelle critique, modification ticket et critique

+++ revu/views.py
@@ -1,5 +1,7 @@
+# from itertools import chain
+from django.db.models import CharField, Value
 from django.shortcuts import render, redirect
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, permission_required
 from django.shortcuts import get_object_or_404
 
 from . import forms, models
@@ -8,27 +10,59 @@ from . import forms, models
 @login_required
 def home(request):
     reviews = models.Review.objects.all()
-    return render(request, 'revu/home.html', context={'reviews': reviews})
+    tickets = models.Ticket.objects.all()
+    context = {
+        'reviews': reviews,
+        'tickets': tickets
+    }
+    return render(request, 'revu/home.html', context=context)
 
 
 def contact(request):
     return render(request, 'revu/contact.html')
 
 
+def no_permission(request):
+    return render(request, 'revu/no_permission.html')
+
+
 @login_required
-def review_upload(request):
+def review_response(request, ticket_id):
+    ticket = get_object_or_404(models.Ticket, id=ticket_id)
     review_form = forms.ReviewForm()
+
     if request.method == 'POST':
         review_form = forms.ReviewForm(request.POST)
         if review_form.is_valid():
             review = review_form.save(commit=False)
+            review.ticket = ticket
+            review.user = request.user
+            review.save()
+            return redirect('home')
+    return render(request, 'review_response.html', {'review_form': review_form, 'ticket': ticket})
+
+
+@login_required
+def create_new_review(request):
+    ticket_form = forms.TicketForm()
+    review_form = forms.ReviewForm()
+    if request.method == 'POST':
+        ticket_form = forms.TicketForm(request.POST)
+        review_form = forms.ReviewForm(request.POST)
+        if all([ticket_form.is_valid(), review_form.is_valid()]):
+            ticket = ticket_form.save(commit=False)
+            ticket.user = request.user
+            ticket.save()
+            review = review_form.save(commit=False)
+            review.ticket = ticket
             review.user = request.user
             review.save()
             return redirect('home')
     context = {
-        'review_form': review_form
+        'ticket_form': ticket_form,
+        'review_form': review_form,
     }
-    return render(request, 'revu/create_review.html', context=context)
+    return render(request, 'revu/create_new_review.html', context=context)
 
 
 @login_required
@@ -40,6 +74,8 @@ def view_review(request, review_id):
 @login_required
 def edit_review(request, review_id):
     review = get_object_or_404(models.Review, id=review_id)
+    if review.user != request.user:
+        return redirect('no_permission')
     edit_form = forms.ReviewForm(instance=review)
     delete_form = forms.DeleteReviewForm()
     if request.method == 'POST':
@@ -48,11 +84,11 @@ def edit_review(request, review_id):
             if edit_form.is_valid():
                 edit_form.save()
                 return redirect('home')
-            if 'delete_review' in request.POST:
-                delete_form = forms.DeleteReviewForm(request.POST)
-                if delete_form.is_valid():
-                    review.delete()
-                    return redirect('home')
+        if 'delete_review' in request.POST:
+            delete_form = forms.DeleteReviewForm(request.POST)
+            if delete_form.is_valid():
+                review.delete()
+                return redirect('home')
     context = {
         'edit_form': edit_form,
         'delete_form': delete_form,
@@ -64,7 +100,7 @@ def edit_review(request, review_id):
 def create_ticket(request):
     ticket_form = forms.TicketForm()
     if request.method == 'POST':
-        ticket_form = forms.TicketForm(request.POST)
+        ticket_form = forms.TicketForm(request.POST, request.FILES)
         if ticket_form.is_valid():
             ticket = ticket_form.save(commit=False)
             ticket.user = request.user
@@ -80,3 +116,53 @@ def create_ticket(request):
 def view_ticket(request, ticket_id):
     ticket = get_object_or_404(models.Ticket, id=ticket_id)
     return render(request, 'revu/view_ticket.html', {'ticket': ticket})
+
+
+@login_required
+def edit_ticket(request, ticket_id):
+    ticket = get_object_or_404(models.Ticket, id=ticket_id)
+    if ticket.user != request.user:
+        return redirect('no_permission')
+    edit_form = forms.TicketForm(instance=ticket)
+    delete_form = forms.DeleteTicketForm()
+    if request.method == 'POST':
+        if 'edit_ticket' in request.POST:
+            edit_form = forms.TicketForm(request.POST, request.FILES, instance=ticket)
+            if edit_form.is_valid():
+                edit_form.save()
+                return redirect('home')
+        if 'delete_ticket' in request.POST:
+            delete_form = forms.DeleteTicketForm(request.POST)
+            if delete_form.is_valid():
+                ticket.delete()
+                return redirect('home')
+    context = {
+        'edit_form': edit_form,
+        'delete_form': delete_form,
+    }
+    return render(request, 'revu/edit_ticket.html', context=context)
+
+
+## Création du flux:
+# def feed(request):
+#     reviews=get_users_viewable_reviews(request.user)
+#     reviews=reviews.annotate(content_type=Value('Review', CharField()))
+
+#     tickets=get_user_viewable_tickets(request.user)
+#     tickets=tickets.annotate(content_type=Value('Ticket', CharField()))
+
+#     posts = sorted(
+#         chain=lambda post: post.time_created,
+#         reverse=True
+#     )
+#     return render(request, 'feed.html', context={'posts': posts})
+
+
+def follow_users(request):
+    form = forms.FollowUsersForm(instance=request.user)
+    if request.method == 'POST':
+        form = forms.FollowUsersForm(request.POST, instance=request.user)
+        if form.is_valid():
+            form.save()
+            return redirect('home')
+    return render(request, 'revu/follow_view.html', context={'form': form})
